delete base_submission_job and move logic into submission since we deleted anonsubmissionjob
